requirements.txt

azureml-core
opencv-python
numpy
pandas
ultralytics
mediapipe
flask
waitress


--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118
torchvision==0.16.2+cu118 


